feat: ShapeDto 생성
